pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wpv"
version = "0.1.0"
description = "Automated water polo video pipeline: 360° footage → ball tracking → reframing → YouTube"
requires-python = ">=3.10"
dependencies = [
    "opencv-python-headless>=4.8",
    "numpy>=1.24",
    "torch>=2.0",
    "torchvision>=0.15",
    "typer>=0.9",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "scipy>=1.10",
    "scikit-image>=0.21",
    "google-api-python-client>=2.90",
    "google-auth-oauthlib>=1.0",
    "flask>=3.0",
]

[project.optional-dependencies]
upscale = ["realesrgan", "basicsr"]
dev = ["pytest>=7.0", "ruff>=0.1"]

[project.scripts]
wpv = "wpv.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/wpv"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100
